// client/src/constants/FormErrors.ts
export enum FormErrorMessage {
  EMPTY_FIELD = 'Ce champs est requis.',
  INVALID_NAME = 'Ce champs doit être un nom valide.',
  INVALID_EMAIL = "Cette adresse email n'est pas valide.",
  INSECURE_PASSWORD = 'Le mot de passe doit contenir au moins 8 caractères dont au moins une majuscule, une minuscule et un chiffre.',
  INVALID_PASSWORD_MATCH = 'Les mots de passe ne correspondent pas.',
}